+++ NEWS.md
@@ -1,4 +1,6 @@
 
+# processx dev
+
 # processx 3.3.1
 
 * Fix a crash on Windows, when a connection that has a pending read