+++ NEWS.md
@@ -5,14 +5,18 @@
   error, if `echo = FALSE`, and it also prints the exit status of the
   process.
 
+* `run()` now includes the standard error in the condition signalled on
+  interrupt.
+
 * `process` now supports creating pseudo terminals on Unix systems.
 
 * `conn_create_pipepair()` gets new argument to set the pipes as blocking
   or non-blocking.
 
 * `process` does not set the inherited extra connections as blocking,
-  and it also does not close them. This now the responsibility of the
-  user. Note that this is a breaking change.
+  and it also does not close them after starting the subprocess.
+  This is now the responsibility of the user. Note that this is a
+  breaking change.
 
 # processx 3.3.1